abstract out common util for verifying validity of a post type ID
