+++ controllers/update-post/index.js
@@ -1,0 +1,2 @@
+const Post = require('../../models/Post')
+const PostType = require('../../models/PostType')
